+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 * Tutorial quests #49
 * `rpg ls` command to look for chests at the current location #51
 * Example sh file #54
-
+* Status effects #48
 ### Changed
 * Tombstones are found with `rpg ls` instead of automatically #52